Only an admin can delete users + Redirect to the main page if a user …

+++ src/Controller/UserController.php
@@ -29,12 +29,15 @@ class UserController extends AbstractController
      */
     public function delete(int $id)
     {
-        // Getting the rowCount value which is returned at the end of deleteTrainerById function
-        $rowCount = $this->userManager->deleteTrainerById($id);
-        $validationMessage = $rowCount == 1 ? 'Le trainer a bien été supprimé!' : 'erreur!';
-        // créer une variable de session $_SESSION['dlt_trainer_msg']
-        $_SESSION['dlt_trainer_msg'] = $validationMessage;
-        header('Location: /User/list');
-        //
+        if (isset($_SESSION['username']) && $_SESSION['userRole'] == "admin") {
+            // Getting the rowCount value which is returned at the end of deleteTrainerById function
+            $rowCount = $this->userManager->deleteTrainerById($id);
+            $validationMessage = $rowCount == 1 ? 'Le trainer a bien été supprimé!' : 'erreur!';
+            // créer une variable de session $_SESSION['dlt_trainer_msg']
+            $_SESSION['dlt_trainer_msg'] = $validationMessage;
+            header('Location: /User/list');
+        } else {
+            header('Location: /');
+        }
     }
 }
